refactor(backend): decompose scanner.py orchestrator (#2364)
Split scanner.py (999 lines) into focused modules:
- indexing_executor.py (64 lines) — thread pool singleton + async dispatch
- file_pipeline.py (335 lines) — file processing pipeline (analysis, storage, iteration)
- indexing_phases.py (226 lines) — ChromaDB init, scan, store orchestration
- subprocess_runner.py (175 lines) — subprocess launch + watchdog supervision

scanner.py reduced to 425-line thin orchestrator with backward-compat re-exports.
All public APIs preserved.
